+++ example/lib/questionnaire_launch_tile.dart
@@ -5,6 +5,7 @@ import 'package:faiadashu_online/restful/restful.dart';
 import 'package:faiadashu_online/url_launch/src/url_launcher.dart';
 import 'package:fhir/r4.dart';
 import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
 
 /// Tile that launches a [QuestionnaireScrollerPage] when tapped.
 class QuestionnaireLaunchTile extends StatefulWidget {
@@ -38,6 +39,9 @@ class QuestionnaireLaunchTile extends StatefulWidget {
 
 class _QuestionnaireLaunchTileState extends State<QuestionnaireLaunchTile> {
   late final FhirResourceProvider _questionnaireProvider;
+  late final Locale _locale;
+  late final NumberFormat _percentPattern;
+  late Future<QuestionnaireModel> _modelFuture;
 
   @override
   void initState() {
@@ -46,13 +50,47 @@ class _QuestionnaireLaunchTileState extends State<QuestionnaireLaunchTile> {
         questionnaireResourceUri, widget.questionnairePath);
   }
 
+  Future<QuestionnaireModel> _createModelFuture() {
+    return QuestionnaireModel.fromFhirResourceBundle(
+            fhirResourceProvider: _questionnaireProvider, locale: _locale)
+        .then<QuestionnaireModel>((qm) {
+      qm.populate(
+          widget.restoreResponseFunction.call(widget.questionnairePath));
+      return qm;
+    });
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _locale = widget.locale ?? Localizations.localeOf(context);
+    _percentPattern = NumberFormat.percentPattern(_locale.toString());
+    _modelFuture = _createModelFuture();
+  }
+
   @override
   Widget build(BuildContext context) {
-    final locale = widget.locale ?? Localizations.localeOf(context);
-
     return ListTile(
       title: Text(widget.title),
-      subtitle: (widget.subtitle != null) ? Text(widget.subtitle!) : null,
+      subtitle: FutureBuilder<QuestionnaireModel>(
+        future: _modelFuture,
+        builder: (context, snapshot) {
+          var countString = '';
+          if (snapshot.hasData) {
+            final _questionnaireModel = snapshot.data!;
+            final _numberCompleted =
+                _questionnaireModel.count((qim) => qim.isAnswered);
+            final _totalNumber =
+                _questionnaireModel.count((qim) => qim.isAnswerable);
+            countString = 'Completed: $_numberCompleted / $_totalNumber '
+                '(${_percentPattern.format(_numberCompleted / _totalNumber)})';
+          }
+          return (widget.subtitle != null)
+              ? Text('${widget.subtitle!}\n$countString')
+              : Text(countString);
+        },
+      ),
+      isThreeLine: true,
       trailing: SizedBox(
         width: 24.0,
         height: 40.0,
@@ -63,7 +101,7 @@ class _QuestionnaireLaunchTileState extends State<QuestionnaireLaunchTile> {
               final questionnaireModel =
                   await QuestionnaireModel.fromFhirResourceBundle(
                       fhirResourceProvider: _questionnaireProvider,
-                      locale: locale,
+                      locale: _locale,
                       aggregators: [NarrativeAggregator()]);
               questionnaireModel.populate(widget.restoreResponseFunction
                   .call(widget.questionnairePath));
@@ -82,7 +120,7 @@ class _QuestionnaireLaunchTileState extends State<QuestionnaireLaunchTile> {
           context,
           MaterialPageRoute(
             builder: (context) => QuestionnaireScrollerPage(
-                locale: locale,
+                locale: _locale,
                 fhirResourceProvider: RegistryFhirResourceProvider([
                   AssetResourceProvider.singleton(
                       questionnaireResourceUri, widget.questionnairePath),
@@ -155,7 +193,12 @@ class _QuestionnaireLaunchTileState extends State<QuestionnaireLaunchTile> {
                   ),
                 ]),
           ),
-        );
+        ).then((value) {
+          // This triggers after return from questionnaire filler
+          setState(() {
+            _modelFuture = _createModelFuture();
+          });
+        });
       },
     );
   }