+++ lib/questionnaires/model/src/questionnaire_model.dart
@@ -302,40 +302,6 @@ class QuestionnaireModel extends QuestionnaireItemModel {
   ///
   int get generation => _generation;
 
-  // TODO: Move this to [QuestionnaireItemModel]?
-
-  /// Returns the index of the first [QuestionnaireItemModel] which matches the predicate function.
-  ///
-  /// The items are examined as returned by [orderedQuestionnaireItemModels].
-  ///
-  /// Returns [notFound] if no matching item exists.
-  int? indexOf(bool Function(QuestionnaireItemModel) predicate,
-      [int? notFound = -1]) {
-    int index = 0;
-    for (final qim in orderedQuestionnaireItemModels()) {
-      if (predicate.call(qim)) {
-        return index;
-      }
-      index++;
-    }
-
-    return notFound;
-  }
-
-  // TODO: Move this to [QuestionnaireItemModel]?
-
-  /// Returns the count of [QuestionnaireItemModel]s which match the predicate function.
-  int count(bool Function(QuestionnaireItemModel) predicate) {
-    int count = 0;
-    for (final qim in orderedQuestionnaireItemModels()) {
-      if (predicate.call(qim)) {
-        count++;
-      }
-    }
-
-    return count;
-  }
-
   /// Returns the [QuestionnaireItemModel] that corresponds to the linkId.
   ///
   /// Throws an [Exception] when no such [QuestionnaireItemModel] exists.